Tournaments improvements
