+++ locales/en-GB.json
@@ -67,17 +67,15 @@
     "TOURNAMENT_CANNOT_SEND_MESSAGE": "I can't send messages in this channel! Please ensure I have the proper permission to do so. Note that while this may look like a message, this is only working because this is a reply to your command.",
     "TOURNAMENT_CREATED": "Tournament created!",
     "CREATE_TOURNAMENT": "Create tournament",
-    "TEAM_SIZE": "Team size:",
     "ATTEMPTS": "Attempts:",
     "DURATION": "Tournament duration:",
     "TOP_FORMAT": "Custom top format:",
     "BOTTON_FORMAT": "Custom bottom format:",
     "TYPE": "Score calculation type:",
-    "TEAM_SIZE_DESCRIPTION": "Default: 1 (solo)",
     "ATTEMPTS_DESCRIPTION": "Default: 3 (three attempts)",
     "DURATION_DESCRIPTION": "Default: 7 (seven days)",
     "TOP_FORMAT_DESCRIPTION": "Default: [{position}] [{player}]",
     "BOTTON_FORMAT_DESCRIPTION": "Default: - {area} ;; {time} ;; {attempt}",
-    "TYPE_DESCRIPTION": "Default: best",
+    "TYPE_DESCRIPTION": "Default: best. Options: best, sum",
     "NOT_A_TOURNAMENT": "This isn't a tournament message!"
 }
