+++ locales/fr.json
@@ -67,13 +67,15 @@
     "TOURNAMENT_CANNOT_SEND_MESSAGE": "Je n'ai pas le permission pour envoyer les messages dedans cette salon!",
     "TOURNAMENT_CREATED": "La tournoi était crée!",
     "CREATE_TOURNAMENT": "Créeir une tournoi:",
-    "TEAM_SIZE": "Grosseur des équipes:",
     "ATTEMPTS": "Nombre des essais:",
     "DURATION": "Durée de tournoi:",
-    "FORMAT": "Format personnalisé:",
-    "TEAM_SIZE_DESCRIPTION": "Défaut: 1 (solo)",
+    "TOP_FORMAT": "Format en haut:",
+    "BOTTON_FORMAT": "Format en bas:",
+    "TYPE": "Score calculation type:",
     "ATTEMPTS_DESCRIPTION": "Défaut: 3 (trois essais)",
     "DURATION_DESCRIPTION": "Défaut: 7 (sept jours)",
-    "FORMAT_DESCRIPTION": "Défaut: [{position}] [{player}]\\n{area} ;; {time} ;; {attempt}",
+    "TOP_FORMAT_DESCRIPTION": "Défaut: [{position}] [{player}]",
+    "BOTTON_FORMAT_DESCRIPTION": "Défaut: - {area} ;; {time} ;; {attempt}",
+    "TYPE_DESCRIPTION": "Défaut: best. Options: best, sum",
     "NOT_A_TOURNAMENT": "Ce n'est pas une message des tournois."
 }
